% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mp_funnel.R
\name{get_mp_funnel}
\alias{get_mp_funnel}
\title{Mixpanel Funnels}
\usage{
get_mp_funnel(api_secret, funnel_id, unit, from_date, to_date)
}
\arguments{
\item{api_secret}{The api secret from your Mixpanel account.}

\item{funnel_id}{The id of the funnel you want to retrieve.}

\item{unit}{The data can be retrieved by day, week, month, or year.}

\item{from_date}{The start date to retrieve the data.}

\item{to_date}{The end date to retrieve the data.}
}
\value{
The funnel steps in order, by date.
}
\description{
This function retrieves funnel data from Mixpanel.
}
