% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mp_funnel_list.R
\name{get_mp_funnel_list}
\alias{get_mp_funnel_list}
\title{Mixpanel Funnels}
\usage{
get_mp_funnel_list(api_secret)
}
\arguments{
\item{api_secret}{The api secret from your Mixpanel account.}
}
\value{
The funnel list.
}
\description{
This function retrieves the funnel list from Mixpanel.
}
